third_party/WebKit/LayoutTests/media/viewport-in-standalone-media-document.html

<!DOCTYPE html>
<title>This tests that a standalone media document has viewport settings.</title>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<iframe></iframe>
<script>
async_test(function(t) {
    var iframe = document.querySelector("iframe");

    iframe.onload = t.step_func_done(function() {
        var metaElement = iframe.contentDocument.querySelector("meta");
        assert_equals(metaElement.name, "viewport");
        assert_equals(metaElement.content, "width=device-width");
    });

    iframe.src = "content/test.ogv";
});
</script>
